update the _SyncRepository endpoint to use graphql
